+++ R/facet-strata.R
@@ -1,6 +1,6 @@
-#' Facet data into n strata
+#' Facet data into groups to facilitate exploration
 #'
-#' @param n_strata Number of facet panels. Default is 12 
+#' @inheritParams stratify_keys
 #' @inheritParams ggplot2::facet_wrap
 #' @import ggplot2
 #'
@@ -8,15 +8,39 @@
 #' @export
 #'
 #' @examples
-#' library(ggplot2)
 #' ggplot(world_heights,
 #'        aes(x = year,
 #'            y = height_cm,
-#'            group = country)) + 
-#'   geom_line() + 
+#'            group = country)) +
+#'   geom_line() +
 #'   facet_strata()
+#' 
+#' ggplot(world_heights,
+#'        aes(x = year,
+#'            y = height_cm,
+#'            group = country)) +
+#'   geom_line() +
+#'   facet_wrap(~continent)
+#' 
+#' ggplot(wages_ts,
+#'        aes(x = xp,
+#'            y = ln_wages,
+#'            group = id)) +
+#'   geom_line() +
+#'   facet_strata(along = xp_since_ged)
+#' 
+#' wages_ts %>%
+#'   key_slope(ln_wages ~ xp) %>%
+#'   right_join(wages_ts, ., by = "id") %>%
+#'   ggplot(aes(x = xp,
+#'              y = ln_wages)) +
+#'   geom_line(aes(group = id)) +
+#'   geom_smooth(method = "lm") + 
+#'   facet_strata(along = .slope_xp)
 
 facet_strata <- function(n_strata = 12,
+                         along = NULL,
+                         fun = mean,
                          nrow = NULL, 
                          ncol = NULL, 
                          scales = "fixed", 
@@ -31,6 +55,8 @@ facet_strata <- function(n_strata = 12,
                       strip.position = strip.position)
   
   facet$params$n <- n_strata
+  facet$params$along <- rlang::enquo(along)
+  facet$params$fun <- fun
   
   ggproto(NULL, 
           FacetStrata,
@@ -53,22 +79,22 @@ FacetStrata <- ggproto(
     
     if (params$as.table) {
       layout$ROW <- as.integer((id - 1L) %/% dims[2] + 1L)
-      } else {
-        layout$ROW <- as.integer(dims[1] - (id - 1L) %/% dims[2])
-      }
+    } else {
+      layout$ROW <- as.integer(dims[1] - (id - 1L) %/% dims[2])
+    }
     
     layout$COL <- as.integer((id - 1L) %% dims[2] + 1L)
-                            
+    
     layout <- layout[order(layout$PANEL), , drop = FALSE]
-
+    
     rownames(layout) <- NULL
     
     # Add scale identification
     layout$SCALE_X <- if (params$free$x) id else 1L
     layout$SCALE_Y <- if (params$free$y) id else 1L
-                            
+    
     cbind(layout, .strata = id)
-    },
+  },
   
   map_data = function(data, 
                       layout, 
@@ -76,10 +102,12 @@ FacetStrata <- ggproto(
     
     if (is.null(data) || nrow(data) == 0) {
       return(cbind(data, PANEL = integer(0)))
-      }
+    }
     
     new_data <- brolgar::stratify_keys(.data = data,
-                                       n_strata = params$n)
+                                       n_strata = params$n,
+                                       along = !!params$along,
+                                       fun = params$fun)
     
     new_data$PANEL = new_data$.strata